Move target models to top level models module
