+++ src/aeonlib/ocs/__init__.py
@@ -1,11 +1,9 @@
 from .request_models import Location, Request, RequestGroup
-from .target_models import Constraints, NonSiderealTarget, SiderealTarget
+from .target_models import Constraints
 
 __all__ = [
     "Location",
     "Request",
     "RequestGroup",
     "Constraints",
-    "SiderealTarget",
-    "NonSiderealTarget",
 ]
