+++ tests/ocs/test_models.py
@@ -3,13 +3,12 @@ from datetime import datetime, timedelta
 import pytest
 from pydantic import ValidationError
 
-from aeonlib.models import Window
+from aeonlib.models import SiderealTarget, Window
 from aeonlib.ocs import (
     Constraints,
     Location,
     Request,
     RequestGroup,
-    SiderealTarget,
 )
 from aeonlib.ocs.lco.instruments import Lco1M0ScicamSinistro
